Rename Layout Files
I had a hard time understanding what each file's purpose was when I
first started
editing them. The words "page" & "section" are kind of conflated. Even
after working with everything I was still having a hard time
remembering the difference between "Index" vs. "SectionIndex". This is
to help clarify a few of these differences for me.
